requirements.txt
Django==1.6.5
Markdown==2.4.1
djangorestframework==2.3.14
pytz==2014.4
psycopg2==2.5.3
